INtentooo

+++ EJ5_RTOS/src/main.c
@@ -80,7 +80,7 @@ static void vTarea3 (void *pvParameters){
 int main(void)
 {
 
-  xQueueCreate(100, sizeo( int ));
+  xQueueCreate(100, sizeo(int long ));
   xQueue
 
 	xTaskCreate(vTarea1, (const char *)"Tarea1", TAM_PILA, (void*)pcTextoTarea1, tskIDLE_PRIORITY+1, NULL );
